project_mercury/rtm/src/main/java/com/smartsparrow/rtm/subscription/courseware/configchange/FeedbackConfigChangeRTMConsumable.java
package com.smartsparrow.rtm.subscription.courseware.configchange;

import com.smartsparrow.rtm.message.RTMClientContext;
import com.smartsparrow.rtm.subscription.courseware.ActivityRTMSubscription;
import com.smartsparrow.rtm.subscription.courseware.message.ConfigChangeBroadcastMessage;
import com.smartsparrow.rtm.subscription.data.AbstractRTMConsumable;
import com.smartsparrow.pubsub.data.RTMEvent;

/**
 * This RTM consumable describes an feedback config change event
 */
public class FeedbackConfigChangeRTMConsumable extends AbstractRTMConsumable<ConfigChangeBroadcastMessage> {

    private static final long serialVersionUID = -8655276725801479183L;

    public FeedbackConfigChangeRTMConsumable(RTMClientContext rtmClientContext, ConfigChangeBroadcastMessage content) {
        super(rtmClientContext, content);
    }

    @Override
    public RTMClientContext getRTMClientContext() {
        return rtmClientContext;
    }

    @Override
    public String getName() {
        return String.format("author.activity/%s/%s", content.getActivityId(), getRTMEvent().getName());
    }

    @Override
    public String getSubscriptionName() {
        return ActivityRTMSubscription.NAME(content.getActivityId());
    }

    @Override
    public RTMEvent getRTMEvent() {
        return new FeedbackConfigChangeRTMEvent();
    }

    @Override
    public boolean equals(Object o) {
        return super.equals(o);
    }

    @Override
    public int hashCode() {
        return super.hashCode();
    }
}
